Moved run_system() from tests/ folder into private FAST-OAD code.
Strangely enough, GitHub workflows have now problems finding tests/testing_utilities.
