+++ src/fastoad/models/geometry/tests/test_aero_center.py
@@ -15,8 +15,8 @@ import os.path as pth
 
 import pytest
 
+from fastoad._utils.testing import run_system
 from fastoad.io import VariableIO
-from tests.testing_utilities import run_system
 from ..compute_aero_center import ComputeAeroCenter
 
 DATA_FOLDER_PATH = pth.join(pth.dirname(__file__), "data")
