+++ src/fastoad/models/handling_qualities/tests/test_static_margin.py
@@ -16,8 +16,8 @@ import os.path as pth
 import openmdao.api as om
 import pytest
 
+from fastoad._utils.testing import run_system
 from fastoad.io import VariableIO
-from tests.testing_utilities import run_system
 from ..compute_static_margin import ComputeStaticMargin
 
 DATA_FOLDER_PATH = pth.join(pth.dirname(__file__), "data")
