+++ src/fastoad/models/weight/cg/tests/test_cg_components.py
@@ -20,8 +20,8 @@ import os.path as pth
 
 import pytest
 
+from fastoad._utils.testing import run_system
 from fastoad.io import VariableIO
-from tests.testing_utilities import run_system
 from ..cg import ComputeAircraftCG
 from ..cg_components.compute_cg_control_surfaces import ComputeControlSurfacesCG
 from ..cg_components.compute_cg_others import ComputeOthersCG
